requirements.txt
opencv-python
ultralytics
kafka-python